Moving frontend, backend connectors and processors into quesma
