+++ quesma/main.go
@@ -16,6 +16,7 @@ import (
 	"quesma/connectors"
 	"quesma/elasticsearch"
 	"quesma/feature"
+	"quesma/frontend_connectors"
 	"quesma/ingest"
 	"quesma/licensing"
 	"quesma/logger"
@@ -29,7 +30,6 @@ import (
 	"quesma/telemetry"
 	"quesma/tracing"
 	"quesma_v2/core"
-	"quesma_v2/frontend_connectors"
 	"runtime"
 	"syscall"
 	"time"
